Formulář pro vložení inzerátu(ADMIN)

+++ src/main/java/org/example/repository/ProductRepository.java
@@ -5,6 +5,7 @@ import org.example.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @Repository
@@ -31,5 +32,5 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
      * @return produkt, pokud existuje
      */
 
-    Product findByPrice(double price);
+    Product findByPrice(BigDecimal price);
 }
